Add files via upload
this is a simple web based implementation of the Rock paper scissors game. It consists of an interactive user interface where one can select their move(could be rock, paper or scissors) and play against the computer that generates random  move. This game tracks and later displays the scores for computer and the user respectively by updating the scores dynamically. It also displays the score of each round. 
The languages used here are HTML, CSS and Javascript