Add files via upload
This project calculates the user's age by taking the date of birth input from the user. I have used HTML, CSS and javascript that serves a specific purpose for the application. For HTML, I have used div with the class container, label, input tags of the type: date for entering the birthdate of the user, button tag to give the output of the calculated age. For CSS, i used the Montserrat font for the body, a container class with gray background and rounded corners, centered on the page. I have used the hover effects for the smoothness and styling on the button. For JS, I have used Event Listener to attach it to the calculateAge function when done onclick. This function gets the birthdate from the user input, then validates it and displays an alert if the area is left empty. If all the inputs are correct, it calculates the age using the getAge function later to get the age. The getAge function uses the current date to calculate the user's age so make sure it has an accurate age calculation. Then the updated age is displayed through #result paragraph with an output indicating the user's age.