Add files via upload
This project is  called a Pomodoro timer count. It is a popular time-based learning strategy that helps the user's to keep themselves productive while taking good amounts of breaks in between ensuring they are not drained or exhausted.

The project consists of HTML, CSS, and JavaScript to implement a countdown timer that focuses on work intervals (typically 25 minutes) followed by short breaks.
The HTML code provides a simple user interface with a display of the time, buttons for starting, stopping and resetting the timer, and added a title to show the purpose of the pagee.
The CSS code provides a simple clean design with distinct color buttons for different actions that is, start, stop and rest and a timer that is displayed on a larger font which makes it easy to read. The buttons also have the hover effect for a smoother experience.
The javascript logic is about the timer that starts with 25minutes which is around 1500 seconds and counts down by updating the timer display everry second. By clicking the start button, the countdown is triggered using setInterval() that uploads the timer every second. The stop button pauses the timer by clearing the time count, which inturn stops the countdown. The reset button stops the timer and resets the countdown back to 25 minutes. The time format used is MM:SS to make sure the countdown displays 2 digits on both sides as it starts off with 25mins and 00 seconds. There is also an alert that is set, as the timer hits zero it notifies the user that the session is completed. 
The main parts of the code is that I have used DOM manipulation by updating the time and responding to the person's actions. I have also used event listenrs by attaching them to the buttons for starting, stopping and resetting the timer enabling better user interactions. Doing this project gave me a strong understanding of functions, event listeners and doms while focused on user interactions and experience