Add files via upload
This project is about a simple stopwatch using html, css and js. The timer begins counting from zero. It updates in real-time, displaying hours, minutes, seconds and milliseconds. When the stop timer is clicked, the timer stops. By clicking the reset button, the timer goes back to 00:00:00. 

The html structure was made simple and includes a display for the time <div id + "timer"> and 3 buttons for starting, stopping and resetting the timer. The buttons also have the id attributes which are start, stop and reset. 
In CSS the styling was minimal with flexbox which was used to center the content horizontally and vertically. The buttons are styled with round edges and larger text and had subtle hover effects. 
In JavaScript, Starting off with the startTimer function that calculates the difference between the currrent time and the start time. The formatTime function formats the time unto a readable format which is hrs: mins: seconds: milliseconds and padded with zeros if necessary. Added event listeners to the buttons for better interaction when the function is executed.