Add files via upload
This project consists of an analog clock which is structured using html and styled with css and functioned with js. 
The html structure structures the page with a clock face in the shape of circle with numbers 12, 3, 6 and 9 placed at appropriate positions and three clock hands namely hours, mins and seconds defined as div inside the ".needle" container.
The css designs the layout of the clock with the face having a light gray background and shadow. The clock hands are styled as div elements with colors and sizes varying from each other. They are then centered using flexbox and the "before" pseudo element shows the pivot of the hand.
In JS, the setinterval function calls the setdate() every second to make sure the hands update continuously. The setdate function also gets the current time using js's date object. It calculates the rotation degrees for the hours, minutes and second hands based on the time. The transform: rotate() property is used to rotate the hands to the calculated degrees of rotation to ensure the hands are exactly where they are supposed to be pointing at.
the hour hand is based on the hour which is gethour/12 multiplied with 360 while minute's hand rotation is getminute/60 multiplied with 360 and lastly seconds hand rotation getsecond/60 multiplied by 360 respectively.