Add files via upload
This project is a simple To-Do List app that lets users add, mark, sort, and remove tasks from a list. I have used html, css and js for this project.

The HTML structure has a container and header where the entire page is wrapped inside class = "container" to keep the layout centered. The inpNewTask id contains an input field where users can type the name of the task they want to add. There are 4 buttons which are added : add, reset, sort and cleanup. The task list id wrapped inside <ul> tag called id = "ulTasks" is used to display the tasks. Each task added to the list is an induvidual <li> element.

In CSS I have used the grid system and the buttons are styled using bootstrap and buttons are designed on the basis of success for add, danger for reset, info for sort, and warning for cleanup.  The task is completed, it will be crossed out and colored with gray shade using the class called "done" which indicates that the task has been completed.

For JS, addItem function is used for adding new tasks where <li> is created whnever the user types a task and clicks the add button.
A click event is added to each list item to toggle the done class. Then the task is appended to the list and then input field is cleared.
The cleanup button with clearDone function removes all tasks that are marked as done from the list. After the cleanup, the button states are re-evaluated by the toggleInputButtons function. The sot button moves completed tasks to the bottom of the list. Add and reset buttons are only enabled if there is text in the input field while sort and cleanup buttons are enabled only if there are tasks in the list and this is managed by the toggleInputButtons function.
I have also used event listeners where when the user presses the enter key while typing the input, a new task is added automatically.  The state of the buttons is updated whenever the user types into the input field, to make sure they can’t add a task with an empty input or reset the field when it’s empty.
The add button calls the addItem() function, reset button clears the input field, sortTasks() is called by the sort button to move the completed tasks to the bottom and cleanup button removes all the completed tasks by calling (cleardone()) function