Add files via upload
This project is about a simple month calendar that displays the current month's dates along with the current date which is made using html, css and js.
In html page, a div element with the class: storage wraps up the entire calender layout to make sure the calender is on the centre of the screen. The calendar itself is is also wrapped in div with the class: calendar.
The month section displays the current month's name and date, the days of the week ssection which labels the days from monday to  sunday and finally the days of the month, which populates based on the current month. 
In the CSS page, some basic styles are applied globally to ensure that the page has no default margin or padding. The font family is set to sans-serif for a clean and modern look. The storage div takes up the full height of the viewport and uses flexbox to center the calendar. 
The month section has a red background with a centered text and hover effects are added to the days (excluding empty slots) for interactivity. The current day is highlighted in a different background color.

In javascript, The script begins by taking the current month and calculating the number of days in the current month (lastday) and the day of the week on which the month starts (firstday). The months array holds the names of the months and later used to display the current month's name dynamically. The monthEL.innertext and fulldateEl.innerText are updated to show the current month and date. There is a loop that runs to create empty div elements before the first day of the month, to align the days correctly on the calendar. The other loop runs to create div elements for each day of the month, marking the current day with a class called "today". The days are then inserted into the DOM, allowing the calendar to reflect the correct month and highlight the current date.