Add files via upload
This program is about rolling a dice that gives a random number from 1-6. In the HTML structure, A heading h1 is used for the project title. A div with the current dice face using unicode characters. A button with an id #roll-button to trigger the dice roll. There is also an unordered list with id #roll-history to keep track of the roll history.
In css, the dice is styled in a larger font for the dice face, so 7rem was used then centered with an animation effect for rolling. The roll-animation uses the keyframes to create the turning or rotating effect when the dice roll button is pressed. The button is styled with a red background, white text and the edges are rounded with a hover effect that changes the bg color upon hovering over. The roll history list is styled with a light background color and slightly shadowed for a modern appearance
In JS, the logic behind the dice roll is that it generates a random number between 1 and 6. It maps the number to a corresponding dice face using the unicode. The unicode is also called a universal character encoding standard. The roll history is stored in an array called "historyList". The animation is handled by added roll-animation class to the dice element on click. The animation class is removed after a timeout to rest the animation back.