Add files via upload
This project is about the password generator that generates a random 14-character password using a combination of numbers, cases and special characters. The password is generated in the input field when the button is clicked. The users can also copy the password to the clipboard by clicking on the copy icon. A green "Password Copied" alert slides in and out smoothly to confirm the action.

The basic html layout includes a header (<h2>) for the title, an input field for displaying the generated password, a copy icon, and a button for generating passwords. There is also an alert container inside the div element with the class "alert-storage" which is used for showing notifications when the password is copied.
In the CSS layout, the body uses the flexbox to center the password generator. The .password-storage div is styled with a fine background, padding and a box shadow to create a card-like appearance. The (.input-storage) is styeld as a flex container to align the input field and copy the icon side by side. The .input for readability with larger font sizing and spacing. The .btn is designed with a black backgound and white text, changing to green on hover and slightly shrinking on active clicks for user feedback. 

In the JS layout, btnEl.addEventListener("click", ...) listens for the clicks to generate a new password once again.  The copyIconEl.addEventListener("click", ...) retrieves the clicks on the copy icon to copy the password and trigger the alert.
When the password is being generated, the createPaaword() function creates a 14-character password by randomly selecting the characters of  alphanumerals and other special chars. 
The copy functionality that is copyPassword() selects the already generated password and copies it to the clipboard using navigator.clipboard.writeText() and the alert container toggles the "active" class which moves it in and out.