Add files via upload
This project is called a live character or letters counter allows users to type into a text area, with real-time updates displaying the total number of characters typed and how many characters remain, based on a limit that has been pre-defined.
The html page contains a container with a heading and a text area where the user can type the information. The area has a max length attribute set to 150 characters as an example. It could be edited anytime to increase or decrease the character count through the code. 
There are also 2 sections just below the text area that display total number of characters typed and number of remaining characters allowed. 
The css page is designed with a layout featuring a centered container that contains the input and counters. It uses a background color of rgb(44,6,44) for the body and 209,213,224 for the container. The text area has a width of 100% and height of 100px and padding is added for better usability. It has a solid border and resize set to none to prevent the user from resizing the text area. 
In the JS, text area element has an event listener attached to it that retrieves the keyup events. Here the update counter() is triggered when the user types or deletes character(s). Then the updatecounter() function helps in updating the total count by getting the text-area value length and calculates the remaining left out characters by subtracting the total from the maxlength attribute. The real-time update of both the counters that is total-counter and remaining-counter gives a clarity that the user is aware of how much space is left in the box.