Add files via upload
The project below is a simple light mode to dark mode switch. It makes the webpage darker when toggled right and light if done to left. 
The html structure consists of an input element which is a checkbox that controls the toggle between dark and light mode. It is hidden from view using visibility: hidden. The label element is used for styling the switch. It can be clicked on and changes the background color and the position of the circle element based on the condition of the checkbox. Speaking of which, <div class="circle"> represents the toggle button that slides from left to right and vice versa to indicate the mode that is being selected.

The css styling is done starting with the body where the background shifts between white for light mode and black for dark mode based on the user's toggle control. The transition is made smooth because of transition: .4s being applied to the body of the code.
The circle is placed in the label and the label is a background for the switch and the circle is animated to move towards left or right based on the state of the checkbox. These are controlled with the toggleOn and Off keyframes which are used to animate the sliding effect. 
In the javascript functionality, the updatebody() function changes the background color of the page depending on whether the checkbox is checked or not. The input event listener checks for changes in the box's state and updates the page's background accordingly. Lastly, the user's theme preference is saved in local storage, provided that the theme is remembered even after the page is reloaded. The theme is stored as a boolean which is true for dark mode and false for light mode.