Add files via upload
This project introduces a Heart-shaped Cursor Effect that adds a playful and interactive visual feature to a webpage. As the user moves their mouse across the screen, the heart-shaped elements are generated at the cursor's position, creating an engaging and dynamic effect.

The html structure is minimal, consisting of a one single body element, which hosts the interactive effects. There are no explicit controls or buttons that is just mouse movement triggers the effect.
Meanwhile on the css page, the heart shaped elements are styled as span tags by using another external image as a background. These elements are put on a random size and placed at the mouse position when the cursor is moved by the user. The heart elements animate using the @keyframes rule. The animation moves the hearts off screen also by adding a smooth motion. The span elements are given the position: absolute and pointer events: none to prevent them from interfering with the other page elements 
In the JS page, the script checks for the mousemove events on the body of the page and as everytime the mouse moves, a new span element is created at the co-ordinates of the mouse that is offsetx and offsety.
Point to note that each new heart element is a given random size and appended to the body. The hearts automatically disappear after 3 seconds using the timeout set called setTimeOut so that there is a good visual experience without cluttering the screen with too many hearts.