Add files via upload
This is a simple calculator that is made using html, css and js. It is very straightforward and allows the user to perform basic arithmetic operations such as as addition, subtraction, multiplication and divison. 
The html coode is responsible for the structure of the calculator as it includes an input field called <input> where the user can see the numbers they have intered and the result after their numbers are calculated. It contains multiple buttons for the numbers (0-9) , operators like +,-,*,/ and additional functionality buttons like c(clear) and =.
These buttons are grouped in a <div> element with the class buttons, and the entire calculator layout is enclosed in a div with the class "calculator".
In the css file, the calculator background is colored black with a padding and a border and centered on the screen. The button styles vary according to their function, namely: .clear has a red background for the clear button, .operator has a blue background for the operator buttons, .equals has a green background and spans three rows to make it stand out, .number and .decimal buttons have a white background for the digit buttons

Finally the js file handles the functionality of the calculator such as, 
the event listeners are added to each button. When the button is clicked, the corresponding value is processed. The clear function that is "clearResult()" evaluates the expresssion in the input field using the eval() function when the '=' button is clicked. The appendvalue() function appends the clicked button's value to the input section.